TestFlight: cloud signing in CI, Xcode 26 build fixes, new app icon
- Workflow signs with an Admin App Store Connect API key (cloud-managed
  certificate), so no .p12 secrets; build numbers are 100 + run number to
  stay above the builds uploaded from a Mac; uses the newest Xcode on the image.
- Fix CGFloat/Double ambiguity errors under Xcode 26 in SchematicRenderer
  and SketchEditing.
- Set the development team, declare no non-exempt encryption, bump build to 5.
- Replace the app icon and commit Package.resolved for SwiftMath.

+++ PhotoMesh/Features/Schematic/SchematicRenderer.swift
@@ -348,7 +348,7 @@ enum SchematicRenderer {
         var previous = CGPoint.zero
         for i in 0...steps {
             let fraction = Double(i) / Double(steps)
-            let angle = clockwise ? startAngle + sweep * fraction : -startAngle - sweep * fraction
+            let angle = CGFloat(clockwise ? startAngle + sweep * fraction : -startAngle - sweep * fraction)
             let point = CGPoint(x: center.x + radius * cos(angle), y: center.y + radius * sin(angle))
             if i == 0 { path.move(to: point) } else { path.addLine(to: point) }
             previous = last
